+++ x-pilot-core/src/main/java/org/xangle/xpilot/core/model/response/TransactionInfo.java
@@ -3,12 +3,14 @@ package org.xangle.xpilot.core.model.response;
 import org.xangle.xpilot.core.entity.TransactionEntity;
 import org.xangle.xpilot.core.service.DateUtilService;
 
+import java.math.BigDecimal;
+
 public record TransactionInfo(
         String transactionHash,
         String age,
         String from,
         String to,
-        Long transactionFee
+        BigDecimal transactionFee
 ) {
     public static TransactionInfo from(TransactionEntity transactionEntity) {
         return new TransactionInfo(
